WIP: support for Airflow 3
